aggregate/requirements.txt
arrow==0.12.1
user-agents==2.0.0
python-dotenv==0.9.1
unidecode==1.0.23
mysql-connector-python==8.0.23
psutil==5.8.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.25.0
google-auth==1.24.0
google-cloud-bigquery==2.6.1
google-cloud-core==1.5.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
idna==2.9
pandas==0.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.9.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
pyarrow==2.0.0
